fix: replace np.float by np.float64

+++ mriqc/interfaces/functional.py
@@ -141,7 +141,7 @@ class FunctionalQC(SimpleInterface):
 
         # FD
         fd_data = np.loadtxt(self.inputs.in_fd, skiprows=1)
-        num_fd = np.float((fd_data > self.inputs.fd_thres).sum())
+        num_fd = np.float64((fd_data > self.inputs.fd_thres).sum())
         self._results["fd"] = {
             "mean": float(fd_data.mean()),
             "num": int(num_fd),
